Things are going to work - importing Go in python
I need to use go >= 1.6beta1, installed with
[gvm](http://www.hostingadvice.com/how-to/install-golang-on-ubuntu/).

After this is simple to build .so file with

    go build -buildmode=c-shared -o sum.so sum.go

For example. As explained
[here](https://blog.filippo.io/building-python-modules-with-go-1-5/)

Note: using go1.5 it does not work.

As descrived in uppon link I tryed also to use resulting sum.h and sum.so to implement a C program main.c for the sum. The file main.c is staged in this commit.

To build the executable:

    gcc -Wall -o main main.c ./sum.so